+++ test/test_ops/test_exp.py
